requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Limiter==1.4
Flask-QRcode==3.0.0
Flask-Stache==0.1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Pillow==8.1.1
pystache==0.5.4
qrcode==6.1
requests==2.25.1
six==1.15.0
urllib3==1.26.3
wallycore==0.8.3
Werkzeug==1.0.1
mysql-connector==2.2.9
